% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/OD_OutputClass.R
\name{as.OD_OccupancyModel}
\alias{as.OD_OccupancyModel}
\title{Create a OD_OccupancyModel object}
\usage{
as.OD_OccupancyModel(data, stan.model, chains, pars)
}
\description{
Create a OD_OccupancyModel object
}

