% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/RN_OutputClass.R
\name{as.RN_OccupancyModel}
\alias{as.RN_OccupancyModel}
\title{Create a RN_OccupancyModel object}
\usage{
as.RN_OccupancyModel(data, stan.model, chains, pars)
}
\description{
Create a RN_OccupancyModel object
}

