% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/OD_OutputClass.R
\name{coef.OD_OccupancyModel}
\alias{coef.OD_OccupancyModel}
\title{Return the model parameter estimates}
\usage{
\method{coef}{OD_OccupancyModel}(object, fun = identity)
}
\description{
Return the model parameter estimates
}

