% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/RN_OutputClass.R
\name{coef.RN_OccupancyModel}
\alias{coef.RN_OccupancyModel}
\title{Return the model parameter estimates}
\usage{
\method{coef}{RN_OccupancyModel}(obj, family = NULL)
}
\description{
Return the model parameter estimates
}

