% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/AuxillaryFunctions.R
\name{get_family}
\alias{get_family}
\title{Return the mcmc chain as a tbl_df object}
\usage{
get_family(object, family = NULL)
}
\description{
Return the mcmc chain as a tbl_df object
}

