% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/AuxillaryFunctions.R
\name{inv.logit}
\alias{inv.logit}
\title{Inverse Logit function}
\usage{
inv.logit(x)
}
\arguments{
\item{x}{a real number}
}
\description{
Inverse Logit function
}
\examples{
inv.logit(3)
}

