% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/AuxillaryFunctions.R
\name{logit}
\alias{logit}
\title{Logit function}
\usage{
logit(p)
}
\arguments{
\item{p}{a probability or proportion}
}
\description{
Calculates the log odds ratio.
}
\examples{
logit(.4)
}

