% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/RN_Analysis.R
\name{makeData.RN}
\alias{makeData.RN}
\title{Create Data using Royle-Nichols Occupancy Model}
\usage{
makeData.RN(Design, Covariate.data = NULL, Abundance.formula = ~1,
  Abundance.params = 0, Detection.formula = ~1, Detection.params = 0)
}
\arguments{
\item{Design}{Data specifying the design of the data. A data frame with columns Species, Site, Day, Observer}

\item{Abundance.formula}{A fixed effects formula for Abundance at each site}

\item{Abundance.params}{Parameters that define Abundance}

\item{Detection.formula}{A fixed effects formula for Detection}

\item{Detection.params}{Parameters that define Detection}
}
\value{
Somthing
}
\description{
Blah blah blah
}

