% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/OD_Analysis.R
\name{Occ.OD}
\alias{Occ.OD}
\title{Analyzes Occupancy data using the Occupancy/Detection model}
\usage{
Occ.OD(Y, Covariate.data = NULL, Occupancy.formula = ~1,
  Detection.formula = ~1, n.chains = 4, n.iter = 1000, fit = NULL,
  inits = NULL)
}
\arguments{
\item{Y}{}

\item{Covariates}{}

\item{Effort}{}

\item{Occupancy}{model}

\item{Occupancy}{params}

\item{Detection}{model}

\item{Detection}{params}
}
\description{
Uses Mackenzie's Occupancy/Detection model.
}

