% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/RN_Analysis.R
\name{Occ.RN.JAGS}
\alias{Occ.RN.JAGS}
\title{Royle-Nichols Occupancy Model using JAGS}
\usage{
Occ.RN.JAGS(Y, Covariate.data = NULL, Abundance.formula = ~1,
  Detection.formula = ~1, n.chains = 4, n.adapt = 1000, n.iter = 1000,
  fit = NA, inits = NULL)
}
\arguments{
\item{Y}{}

\item{Covariate.data}{}
}
\description{
Royle-Nichols Occupancy Model using JAGS
}

