% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/RN_Analysis.R
\name{Occ.RN}
\alias{Occ.RN}
\title{Royle-Nichols Occupancy Model}
\usage{
Occ.RN(Y, Covariate.data = NULL, Abundance.formula = ~1,
  Detection.formula = ~1, N.limit = 10, n.chains = 4, n.adapt = 2000,
  n.iter = 2000, num.cores = 1, fit = NA, inits = NULL)
}
\arguments{
\item{Y}{A data frame with 5 columns - Species, Site, Observer, Effort, NumDetections}

\item{Abundance.formula}{Model formula specifying the relationship between Abundance and the covariates}

\item{Detection.formula}{Model formula specifying the relationship between Detection probability and the covariates}

\item{N.limit}{Maximum conceivable number of individuals at a site}

\item{n.chains}{}

\item{n.adapt}{}

\item{n.iter}{}

\item{fit}{}

\item{inits}{}

\item{covariate.data}{Data frame with the same number of rows as Y}

\item{Species}{If there are multiple species, what column in the covariate data frame indicates species.}
}
\value{
An object of type RN_OccupancyModel
}
\description{
Blah blah some description of the model
}

