% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/OD_Analysis.R
\name{OD.beta2Psi}
\alias{OD.beta2Psi}
\title{Mckenzie Model convert beta to psi}
\usage{
OD.beta2Psi(beta, X)
}
\arguments{
\item{beta}{A vector of length p}

\item{X}{A design matrix of dimensions n-by-p}
}
\description{
Mckenzie Model convert beta to psi
}

