% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/OD_Analysis.R
\name{OD.beta2r}
\alias{OD.beta2r}
\title{McKenzie Model convert alpha to r}
\usage{
OD.beta2r(beta, W)
}
\arguments{
\item{W}{A design matrix of dimensions n-by-p}

\item{alpha}{A vector of length p}
}
\description{
McKenzie Model convert alpha to r
}

