% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/OD_Analysis.R
\name{OD.Psi2beta}
\alias{OD.Psi2beta}
\title{Mckenzie Model convert psi to beta}
\usage{
OD.Psi2beta(Psi, X)
}
\arguments{
\item{Psi}{A vector of length n}

\item{X}{A design matrix of dimensions n-by-p}
}
\description{
Mckenzie Model convert psi to beta
}

