% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/OD_Analysis.R
\name{OD.r2beta}
\alias{OD.r2beta}
\title{McKenzie Model convert r to alpha}
\usage{
OD.r2beta(r, W)
}
\arguments{
\item{r}{A vector of length n}

\item{W}{A design matrix of dimensions n-by-p}
}
\description{
McKenzie Model convert r to alpha
}

