% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/RN_Analysis.R
\name{RN.alpha2r}
\alias{RN.alpha2r}
\title{Royle-Nichols Model convert alpha to r}
\usage{
RN.alpha2r(alpha, W)
}
\arguments{
\item{alpha}{A vector of length p}

\item{W}{A design matrix of dimensions n-by-p}
}
\description{
Royle-Nichols Model convert alpha to r
}

