% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/RN_Analysis.R
\name{RN.beta2Psi}
\alias{RN.beta2Psi}
\title{Royle-Nichols Model convert beta to psi}
\usage{
RN.beta2Psi(beta, X)
}
\arguments{
\item{beta}{A vector of length p}

\item{X}{A design matrix of dimensions n-by-p}
}
\description{
Royle-Nichols Model convert beta to psi
}

