% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/RN_Analysis.R
\name{RN.Psi2beta}
\alias{RN.Psi2beta}
\title{Royle-Nichols Model convert psi to beta}
\usage{
RN.Psi2beta(Psi, X)
}
\arguments{
\item{Psi}{A vector of length n}

\item{X}{A design matrix of dimensions n-by-p}
}
\description{
Royle-Nichols Model convert psi to beta
}

