% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/RN_Analysis.R
\name{RN.r2alpha}
\alias{RN.r2alpha}
\title{Royle-Nichols Model convert r to alpha}
\usage{
RN.r2alpha(r, W)
}
\arguments{
\item{r}{A vector of length n}

\item{W}{A design matrix of dimensions n-by-p}
}
\description{
Royle-Nichols Model convert r to alpha
}

