% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/AuxillaryFunctions.R
\name{safe_mean}
\alias{safe_mean}
\title{Safe Averaging}
\usage{
safe_mean(x)
}
\arguments{
\item{x}{a vector of numerical, logical, or factors}
}
\description{
A function that calculates the mean if given numerical values.
If passed a vector of factors, it checks if the levels are all the same
and returns that level, otherwise it can't average factor levels and
errors out.
}
\examples{
safe_mean( c(1,2,3) )
safe_mean( factor('A','A','A') )
# Not run, should throw an error.
# safe_mean( factor('A','B','B') )
}

