% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/AuxillaryFunctions.R
\name{toRadianTime}
\alias{toRadianTime}
\title{Convert Time from 0-24:00 to 0-2*pi}
\usage{
toRadianTime(x)
}
\arguments{
\item{x}{A POSIXct date/time object}
}
\description{
Because the package overlap parameterizes time of day
from 0 to 2*pi, times need to be converted.
}
\examples{
x <- strptime('09-18-2010 17:30:05', format='\%m-\%d-\%Y \%H:\%M:\%S')
toRadianTime(x)

# You can also use package lubridate to create date/time objects
library(lubridate)
y <- mdy_hms('09-18-2010 06:00:05')
toRadianTime(y)
}

